Widen struct member types in bitwise_ops_struct_members.c
c6502 narrow types broke three of the test's five checks:

* `unsigned long l = 2147483650ul` — c6502's `unsigned long` is
  2 bytes (max 65535), so 2147483650 truncates to 2 and check 1
  (`i.b | o.l` expecting 2147483747) gets `97 | 2 = 99`.
* `unsigned int u = 100000u` — c6502's `unsigned int` is 1 byte
  (max 255), so 100000 truncates to 160 and check 2
  (`o.bar ^ i.u` expecting 100036) gets `100 ^ 160 = 196`.
* `i.b << o.bar` with `o.bar = 12` — `i.b` is char which
  integer-promotes to c6502's 1-byte `int`. Per C99 §6.5.7.3
  the shift result type is the promoted left's type, so the
  shift happens at 1-byte width and `97 << 12` is UB.

Adaptation:
* `unsigned long l` → `unsigned long long l` (4 bytes).
* `unsigned int u` → `unsigned long long u` (4 bytes — bitwise
  ops just need the bits to survive, the unsigned-int-specific
  semantics aren't being tested here).
* Check 5: cast `i.b` to `long long` so the shift's result type
  is 4-byte, fitting `97 << 12 = 397312`.

Pin updated 1 → 0.

+++ tests/chapter_18/valid/extra_credit/other_features/bitwise_ops_struct_members.c
@@ -1,12 +1,24 @@
 // Bitwise operations with structure members
+//
+// c6502 adaptation:
+//   * `unsigned long l` widened to `unsigned long long l` so
+//     2147483650 (= 2^31 + 2) fits — c6502's `unsigned long` is
+//     2 bytes (max 65535) and would truncate to 2.
+//   * `unsigned int u` widened to `unsigned long long u` so
+//     100000 fits — c6502's `unsigned int` is 1 byte (max 255)
+//     and would truncate to 160.
+//   * Check 5's shift `i.b << o.bar` casts `i.b` to `long long`:
+//     the promoted-left result type per §6.5.7.3 is then 4-byte
+//     instead of 1-byte, so 97 << 12 = 397312 doesn't overflow
+//     the shift width or wrap.
 
 struct inner {
     char b;
-    unsigned int u;
+    unsigned long long u;
 };
 
 struct outer {
-    unsigned long l;
+    unsigned long long l;
     struct inner *in_ptr;
     int bar;
     struct inner in;
@@ -33,7 +45,7 @@ int main(void) {
     }
 
     o.bar = 12;
-    if ((i.b << o.bar) != 397312) {
+    if (((long long)i.b << o.bar) != 397312) {
         return 5;
     }
